# server/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Local LLM Bridge API
  version: 1.0.0
  description: OpenAI-compatible bridge with agent loop, tools, sessions, jobs, and RAG.
paths:
...
  /v1/task-types/{key}/workflow:
    get:
      summary: Read a task type's ordered workflow stages
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
    put:
      summary: Replace a task type's whole workflow stage set
      description: >-
        422 {error, problems:[{column_slug, field, message}]} on an unknown
        behaviour key, a behaviour attached at the wrong scope, an invalid
        param, or more than one worker/approver participant in a stage.
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The saved stage set
        '422':
          description: Validation problems
  /v1/workflows:
    get:
      summary: List every task type's full workflow (stages + behaviours)
      security:
        - bearerAuth: []
...
  /v1/agents/{agentId}/subscriptions:
    get:
      summary: An agent's column subscriptions, with each column's task-type filter
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
    put:
      summary: Replace an agent's column subscriptions
      description: >-
        Accepts subscriptions:[{column_slug, task_types}] (task_types null or
        omitted = every type); the legacy column_slugs:[...] body is still
        accepted and saves each column with no filter.
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Saved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...